requirements.txt
MySQL-python
mysqlclient 
python-dotenv
openpyxl